+++ Src/Library/changelog.md
@@ -1,2 +1,5 @@
+### NEW
+- support for `IParsable<T>` interface from .net 7.0 #385
+
 ### IMPROVEMENTS
 - testing extensions now return a `TestResult<TResponse>` record instead of a value tuple